+++ app/__init__.py
@@ -3,10 +3,10 @@ from flask_restful import Resource, Api
 
 # local import
 from instance.config import app_config
-from .api.v1.views import redflag as rf
-from .api.v1.views import incident_api
-from .api.v1.views.redflag_view import RedFlags, RedFlag
-from .api.v1.views.incident_view import IncidentApi
+from .v1.views import redflag as rf
+from .v1.views import incident_api
+from .v1.views.redflag_view import RedFlags, RedFlag
+from .v1.views.incident_view import IncidentApi
 
 def create_app(config_name):
     app = Flask(__name__, instance_relative_config=True)
